feat(spells): unify spell rows + refine free-casting UI (OSC-107)
- Merge FreeCastRow + SpellCastRow into one data-driven SpellRow so cast/spent
  styling lives in one place; cast button now uses the ui/Button primitive.
- SpellLevelBadge: shared brass "L{n}" badge (IM Fell), shown after the spell
  name; no strike-through when spent.
- PillSelect level tabs in IM Fell; slot count as a circular badge (inactive
  recedes, active stands out + brass number — token-only, theme-adaptive).
- Free-casting level head pips show uses-left (was spent).
- Spell summaries show the roll formula: read the real `system.roll` field
  (OSE has no `damage` field); drop the bogus type field; rename meta kind.
- Cast button holds its width so the spinner state can't shrink it.
