+++ src/OscSheet/features/actions/MemorizedSpells.tsx
@@ -1,8 +1,7 @@
 import type { OSEActor, OseSpell } from "@domain/types";
 import { useOscSheetContext } from "@app/context";
 import { SectionTitle } from "@ui/SectionTitle";
-import { SpellCastRow } from "@features/spells/SpellCastRow";
-import { FreeCastRow } from "@features/spells/FreeCastRow";
+import { SpellRow } from "@features/spells/SpellRow";
 import {
   spellMeta,
   memorizationDisabled,
@@ -16,7 +15,7 @@ type Props = { actor: OSEActor };
 
 /**
  * Quick-cast list on the Actions tab. With memorization ENABLED (default) it
- * lists the selected/memorized spells (rich <SpellCastRow> with cast dots).
+ * lists the selected/memorized spells (a <SpellRow> with the level badge).
  * With memorization DISABLED it lists FAVORITED spells instead, cast against
  * each level's shared spell-point budget.
  */
@@ -24,7 +23,7 @@ export function MemorizedSpells({ actor }: Props) {
   const { canEdit } = useOscSheetContext();
   const meta = (spell: OseSpell) =>
     spellMeta(spell).map((p) => (
-      <span key={p.kind} className={cx(p.kind === "damage" && "dmg")}>
+      <span key={p.kind} className={cx(p.kind === "roll" && "dmg")}>
         {p.text}
       </span>
     ));
@@ -42,13 +41,15 @@ export function MemorizedSpells({ actor }: Props) {
             const max = (slots[lvl] ?? { max: 0 }).max;
             const left = pointsLeftAt(actor, lvl, max);
             return (
-              <FreeCastRow
+              <SpellRow
                 key={spell._id as string}
                 spell={spell}
                 meta={meta(spell)}
+                free
                 levelTag={lvl}
                 pool={{ used: max - left, max }}
-                exhausted={left <= 0}
+                spent={left <= 0}
+                spentTitle="No spell points left at this level (Study to recover)"
                 canCast={canEdit}
                 onCast={() => castFree(actor, spell, max)}
                 onOpenName={() => spell.sheet.render(true)}
@@ -73,11 +74,13 @@ export function MemorizedSpells({ actor }: Props) {
       <SectionTitle hint="click to cast">Memorized Spells</SectionTitle>
       <div className="fvtt-castlist">
         {spells.map((spell) => (
-          <SpellCastRow
+          <SpellRow
             key={spell._id as string}
             spell={spell}
-            rowClass="osc-spell"
             meta={meta(spell)}
+            levelTag={spell.system.lvl}
+            spent={(spell.system.cast ?? 0) <= 0}
+            spentTitle={`${spell.name} — spent (Rest to recover)`}
             canCast={canEdit}
             onCast={() => spell.spendSpell({ skipDialog: false })}
             onOpenName={() => spell.sheet.render(true)}
