+++ src/OscSheet/features/spells/spells.test.ts
@@ -14,19 +14,19 @@ const spell = (system: Partial<OseSpell["system"]>) =>
   ({ system }) as OseSpell;
 
 describe("spellMeta", () => {
-  it("orders range · duration · save · damage and prefixes R/D", () => {
+  it("orders range · duration · save · roll and prefixes R/D", () => {
     const parts = spellMeta(
-      spell({ range: "150'", duration: "1 turn", save: "vs spells", damage: "1d6+1" })
+      spell({ range: "150'", duration: "1 turn", save: "vs spells", roll: "1d6+1" })
     );
     expect(parts).toEqual([
       { kind: "range", text: "R 150'" },
       { kind: "duration", text: "D 1 turn" },
       { kind: "save", text: "save vs spells" },
-      { kind: "damage", text: "1d6+1" },
+      { kind: "roll", text: "1d6+1" },
     ]);
   });
 
-  it("renders 'no save' when there is no save, and drops empty range/duration/damage", () => {
+  it("renders 'no save' when there is no save, and drops empty range/duration/roll", () => {
     const parts = spellMeta(spell({ range: "", duration: "", save: "" }));
     expect(parts).toEqual([{ kind: "save", text: "no save" }]);
   });
